+++ src/main/kotlin/dev/silenium/multimedia/util/Errors.kt
@@ -3,7 +3,7 @@ package dev.silenium.multimedia.util
 class AVException(val operation: String, val error: Int) :
     Exception("FFmpeg error during $operation: ${error.asAVErrorString()}")
 
-class EGLException(val operation: String, val error: Int) :
+class EGLException(val operation: String, val error: Long) :
     Exception("EGL error during $operation: ${error.asEGLErrorString()}")
 
 class GLException(val operation: String, val error: Int) :
@@ -14,14 +14,14 @@ class VAException(val operation: String, val error: Int) :
 
 fun Int.asAVError(operation: String = "ffmpeg call"): Exception = AVException(operation, this)
 fun Int.asAVErrorString(): String = avErrorStringN(this)
-fun Int.asEGLError(operation: String = "EGL call"): Exception = EGLException(operation, this)
-fun Int.asEGLErrorString(): String = eglErrorStringN(this)
+fun Long.asEGLError(operation: String = "EGL call"): Exception = EGLException(operation, this)
+fun Long.asEGLErrorString(): String = eglErrorStringN(this)
 fun Int.asGLError(operation: String = "GL call"): Exception = GLException(operation, this)
 fun Int.asGLErrorString(): String = glErrorStringN(this)
 fun Int.asVAError(operation: String = "VA call"): Exception = VAException(operation, this)
 fun Int.asVAErrorString(): String = vaErrorStringN(this)
 
 private external fun avErrorStringN(error: Int): String
-private external fun eglErrorStringN(error: Int): String
+private external fun eglErrorStringN(error: Long): String
 private external fun glErrorStringN(error: Int): String
 private external fun vaErrorStringN(error: Int): String
